stored: move DeviceControlRecord to a separate header file
