+++ core/src/plugins/stored/autoxflate-sd.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2013-2014 Planets Communications B.V.
-   Copyright (C) 2013-2016 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -28,6 +28,7 @@
  */
 #include "include/bareos.h"
 #include "stored/stored.h"
+#include "stored/device_control_record.h"
 
 #if defined(HAVE_LIBZ)
 #include <zlib.h>
