+++ core/src/plugins/stored/scsicrypto-sd.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2012 Planets Communications B.V.
-   Copyright (C) 2013-2016 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -60,6 +60,7 @@
  * libbareos shared library.
  */
 #include "include/bareos.h"
+#include "stored/device_control_record.h"
 #include "stored/device_status_information.h"
 #include "stored/jcr_private.h"
 #include "stored/stored.h"
