+++ core/src/stored/acquire.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2002-2013 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2018 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -33,6 +33,7 @@
 #include "stored/acquire.h"
 #include "stored/autochanger.h"
 #include "stored/bsr.h"
+#include "stored/device_control_record.h"
 #include "stored/job.h"
 #include "stored/label.h"
 #include "stored/sd_plugins.h"
