+++ core/src/stored/ansi_label.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2005-2009 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2016 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -29,7 +29,8 @@
  */
 
 #include "include/bareos.h" /* pull in global headers */
-#include "stored/stored.h"  /* pull in Storage Daemon headers */
+#include "stored/device_control_record.h"
+#include "stored/stored.h" /* pull in Storage Daemon headers */
 #include "stored/stored_globals.h"
 #include "stored/label.h"
 #include "stored/ebcdic.h"
