+++ core/src/stored/backends/generic_tape_device.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2014-2014 Planets Communications B.V.
-   Copyright (C) 2014-2014 Bareos GmbH & Co. KG
+   Copyright (C) 2014-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -31,6 +31,7 @@
  */
 
 #include "include/bareos.h"
+#include "stored/device_control_record.h"
 #include "stored/stored.h"
 #include "generic_tape_device.h"
 #include "stored/autochanger.h"
