+++ core/src/stored/bextract.cc
@@ -34,6 +34,7 @@
 #include "lib/crypto_cache.h"
 #include "stored/acquire.h"
 #include "stored/butil.h"
+#include "stored/device_control_record.h"
 #include "stored/jcr_private.h"
 #include "stored/mount.h"
 #include "stored/read_record.h"
