+++ core/src/stored/bsr.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2002-2010 Free Software Foundation Europe e.V.
-   Copyright (C) 2016-2016 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -40,6 +40,7 @@
 
 #include "include/bareos.h"
 #include "stored/bsr.h"
+#include "stored/device_control_record.h"
 #include "stored/jcr_private.h"
 #include "stored/stored.h"
 #include "include/jcr.h"
