+++ core/src/stored/dev.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2019 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -77,6 +77,7 @@
 #include "stored/stored.h"
 #include "stored/autochanger.h"
 #include "stored/bsr.h"
+#include "stored/device_control_record.h"
 #include "stored/jcr_private.h"
 #include "stored/sd_backends.h"
 #include "lib/btimers.h"
