+++ core/src/stored/record.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2001-2012 Free Software Foundation Europe e.V.
-   Copyright (C) 2016-2019 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -31,6 +31,7 @@
 #include "include/bareos.h"
 #include "stored/jcr_private.h"
 #include "stored/stored.h"
+#include "stored/device_control_record.h"
 #include "lib/attribs.h"
 #include "lib/util.h"
 #include "include/jcr.h"
@@ -685,8 +686,7 @@ bool DeviceControlRecord::WriteRecord()
   }
 
   jcr->JobBytes += after_rec->data_len; /* increment bytes this job */
-  if (jcr->impl->RemainingQuota &&
-      jcr->JobBytes > jcr->impl->RemainingQuota) {
+  if (jcr->impl->RemainingQuota && jcr->JobBytes > jcr->impl->RemainingQuota) {
     Jmsg0(jcr, M_FATAL, 0, _("Quota Exceeded. Job Terminated.\n"));
     goto bail_out;
   }
