+++ core/src/win32/stored/backends/win32_fifo_device.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2013-2013 Planets Communications B.V.
-   Copyright (C) 2013-2013 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -32,6 +32,7 @@
 
 #include "include/bareos.h"
 #include "lib/berrno.h"
+#include "stored/device_control_record.h"
 #include "stored/stored.h"
 #include "win32_fifo_device.h"
 #include "stored/autochanger.h"
